requirements.txt
validators==0.12.1
requests>=2.18.4
python-slugify>=1.2.5
